changed the print function for the button and Lights class

+++ Bye-Cycle_FollowingSystem/ldr.h
@@ -8,6 +8,6 @@
 //
 //@param analogInput : This is the pin where the analog sensor is wired to
 //
-long returnAverage(byte analogInput);
+int returnAverage(byte analogInput);
 
 #endif
